Fixed issue with AI switching in fainted mons: Changed variable u8 -> u16 so pokemon with speeds of 256 (and above) won't overflow.

+++ src/battle_ai_switch_items.c
@@ -1062,7 +1062,7 @@ u8 GetMostSuitableMonToSwitchInto(void)
     s32 i, j;
     u8 invalidMons;
     u16 move;
-    u8 consideredSpeed, bestSpeed;
+    u16 consideredSpeed, bestSpeed;
 
     if (*(gBattleStruct->monToSwitchIntoId + gActiveBattler) != PARTY_SIZE)
         return *(gBattleStruct->monToSwitchIntoId + gActiveBattler);
